+++ services/normalization/index.js
@@ -1,0 +1,6 @@
+const { normalizeProduct, normalizeProducts } = require('./product-normalizer');
+
+module.exports = {
+  normalizeProduct,
+  normalizeProducts,
+};